refactor(test): move mock files to src/mock directory
- Relocate testUtils.js to src/mock/testMock.js
- Improve test file organization
