+++ __tests__/SuccessTest.js
@@ -1,9 +1,5 @@
 import App from '../src/App.js';
-import {
-  getLogSpy,
-  mockQuestions,
-  mockRandoms,
-} from '../src/utils/testUtils.js';
+import { getLogSpy, mockQuestions, mockRandoms } from '../src/mock/testMock.js';
 
 const MOVING_FORWARD = 4;
 const STOP = 3;
